+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 This project does not follow SemVer, since modules are independent of each other; thus, SemVer does not make sense. Changes are grouped per module.
 
 ## [Unreleased]
-### script/database_role
+### script/database-roles
 - Add SQL script for creating DB readonly role [#177](https://github.com/dbl-works/terraform/pull/177)
 
 ## [v2023.03.06]